rename and shorten flags
